+++ src/main/java/com/example/floud/dto/request/user/LoginRequestDto.java
@@ -1,4 +1,4 @@
-package com.example.floud.dto.request;
+package com.example.floud.dto.request.user;
 
 import lombok.*;
